+++ src/modules/append_no_of_likes.js
@@ -1,12 +1,10 @@
 const appendNoOFLike = (id, likeNumber) => {
-  const likeHtml = document.querySelector(`#likes-${id}`);
-  likeHtml.innerHTML = `${likeNumber} likes`;
+  document.querySelector(`#likes-${id}`)
+    .innerHTML = `${likeNumber} likes`;
 };
 
 const appendNoOFLikes = (likesArray = []) => {
-  const likeNodes = document.querySelectorAll('.like');
-
-  likeNodes.forEach((node) => {
+  document.querySelectorAll('.like').forEach((node) => {
     const id = node.id.match(/\d+/)[0];
     const { likes } = likesArray.find((obj) => obj.item_id === id);
     appendNoOFLike(id, likes);
