+++ src/modules/display_list_items.js
@@ -1,6 +1,6 @@
 import likeListener from './like_listener';
 import Api from './api';
-import appendNoOFLikes from './append_of_likes';
+import appendNoOFLikes from './append_no_of_likes';
 import { appendCount } from './counter';
 import { appendListItem } from './create_listhtml';
 import { popupInit } from './popup';
@@ -14,7 +14,6 @@ export const displayItems = async () => {
   likeListener();
   const likes = await api.getLikes();
   appendNoOFLikes(likes);
-  
 };
 
 displayItems();
