+++ src/modules/like_listener.js
@@ -1,5 +1,5 @@
 import Api from './api';
-import appendNoOFLikes from './append_of_likes';
+import appendNoOFLikes from './append_no_of_likes';
 
 const addLikeListener = (heart, id) => {
   heart.addEventListener('click', async () => {
